organize navbar + footer
