+++ src/app/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from "next";
 import { Inter, Plus_Jakarta_Sans } from "next/font/google";
 import "./globals.css";
-import Navbar from "@/components/Navbar";
-import Footer from "@/components/Footer";
+import Navbar from "@/components/navigations/Navbar";
+import Footer from "@/components/navigations/Footer";
 
 const inter = Inter({ subsets: ["latin"], fallback: ["sans-serif"] });
 const jakarta = Plus_Jakarta_Sans({
